Started Final Project
Created monopolyFinalProject directory to store all the project files.
Configured the project to SFML 2.5.1
Implemented the intro screen with play button.
Implemented the selection screen to make a choice for "New Game" or "Load Game".
Implemented the player number selection screen that allows to select how many players will be in the game.