Bug Fixes in Dice
Previously, dice was moved the player after he had played
all his turns rather than waiting and performing the specified
action. Fixed that. Now the dice, waits for the specified action
to complete so that the game continues according to the rules.

Added the functionality to send the player to jail when he rolls
three doubles.

Next up, need to add the functionality of bringing out the player
when he rolls double while in jail.

Once the Player Class are completed, the code will be integrated
into the remaining components.